agent1_gesture/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
onnxruntime==1.18.0
numpy==1.26.4
python-dotenv==1.0.1
